+++ workspaces/tekton/plugins/tekton/CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Dependencies
 
+## 3.27.5
+
+### Patch Changes
+
+- 56b4264: use `usek8sobjects` hook from k8s-react package
+
+## 3.27.4
+
+### Patch Changes
+
+- 4523634: Updated dependency `@testing-library/jest-dom` to `6.7.0`.
+
 ## 3.27.3
 
 ### Patch Changes